Updates case data: 2020-04-07. Minor format update (testing_cumulative.csv).
In testing_cumulative, column "testing_cumulative" is now purely integer. Extra info (previously a "*" after the integer value) has been moved to a new column, "testing_info". This should not break any existing code, since reading the integer column in as a string, stripping it off any "*" values, and converting to integer should be harmless for a column that is already purely integer.