New data: 2020-07-31; revise historical data. Note that QC case numbers today may differ from other sources.
Revise historical data: cases (ON, SK).

QC has been transitioning to a new data platform, which has resulted in an inability to provide recovered numbers for several days.

QC has also revised case numbers down in some regions as a result of their new data platform. In keeping with our usual practice when the dates of cases to remove cannot be identified, we have removed recent cases in order to ensure health region cumulative case counts are correct as of today.

Data from some ON PHUs may be temporarily disrupted in the near future due to switching over to a new data platform.