New data: 2020-08-07; revise historical data.
Revise historical data: cases (ON, SK).

ON - Peel data dashboard has been updated recently; we have updated the time series for this health region to align with it.

QC has been transitioning to a new data platform, which has resulted in an inability to provide recovered numbers for some time.