New data: 2020-09-09; revise historical data.
Revise historical data: cases (AB, BC, ON, SK).

Simcoe Muskoka (ON): Regarding yesterday’s (2020-09-08) date note, we have re-assessed the update to be more in line with our usual policies. The daily increase was announced to be 17 cases with 2 previous cases removed, for a net increase of +15. Contrary to the note yesterday, we did not keep the old date (2020-09-03) on the 2 removed cases, rather the new date was assigned (2020-09-08). Thus, the new cases in the health region time series yesterday is +17 rather than +15.