New data: 2020-09-14; revise historical data. Major rev. of SK cases.
Revise historical data: cases (BC, QC, SK); testing (AB).

SK transitioned to reporting according to a new, expanded set of health regions today. These regions do not align exactly with the old health regions. However, the time series with the new health regions provided on their website goes back only to August 4, making a full update of the time series impossible at this time.

At the moment, we are doing our best to re-map cases from the new health regions back onto the old health regions in order to provide a consistent time series. We know this method will produce some inaccuracies, which we hope we will be able to rectify in the future. We have reached out to the SK government to discuss how this issue can be resolved.

See the following link for a comparison of the old and new datasets provided by the SK government (https://dashboard.saskatchewan.ca/health-wellness/covid-19/cases):

https://github.com/jeanpaulrsoucy/covid-19-canada-gov-data/tree/master/sk/cases-by-region