New data: 2020-09-27; revise historical data. See note re: SK data.
Revise historical data: cases (MB, ON, QC).

Although Yukon detected their first COVID case in 50 days, the case will not be included in our dataset. “Since the person is not a resident of Yukon, the release noted that the positive infection will not be included in Yukon's COVID-19 case count.” This is consistent with the behaviour of other provinces.

https://www.ctvnews.ca/health/coronavirus/yukon-detects-first-covid-19-case-in-50-days-1.5122085

Note about SK data: SK transitioned to reporting according to a new, expanded set of health regions on 2020-09-14. Unfortunately, the new health regions do not correspond exactly to the old health regions. Additionally, the provided case time series using the new boundaries do not exist for dates earlier than August 4, making providing a time series using the new boundaries impossible.

For now, we are adding new cases according to the list of new cases given in the “highlights” section of the SK government website (https://dashboard.saskatchewan.ca/health-wellness/covid-19/cases). These new cases are roughly grouped according to the old boundaries. However, health region totals were redistributed when the new boundaries were instituted on 2020-09-14, so while our daily case numbers match the numbers given in this section, our cumulative totals do not. We have reached out to the SK government to determine how this issue can be resolved. We will rectify our SK health region time series as soon it becomes possible to do so.